Renaming and moving properties to view model
