+++ WooCommerce/WooCommerceTests/ViewRelated/Shipping Label/WooShipping Create Shipping Labels/WooShippingAddCustomPackageViewModel.swift
@@ -2,20 +2,23 @@ import XCTest
 @testable import WooCommerce
 import Yosemite
 
-final class WooShippingAddPackageViewModelTests: XCTestCase {
+final class WooShippingAddCustomPackageViewModelTests: XCTestCase {
     func test_it_inits_with_empty_field_values() {
         // Given/When
-        let viewModel = WooShippingAddPackageViewModel()
+        let viewModel = WooShippingAddCustomPackageViewModel()
 
         // Then
         XCTAssertNotNil(viewModel)
         XCTAssertEqual(viewModel.fieldValues.isEmpty, true)
+        XCTAssertEqual(viewModel.packageType, WooShippingPackageType.box)
+        XCTAssertEqual(viewModel.showSaveTemplate, false)
+        XCTAssertEqual(viewModel.packageTemplateName, "")
         XCTAssertEqual(viewModel.areFieldValuesInvalid, true)
     }
 
     func test_clear_field_values() {
         // Given
-        let viewModel = WooShippingAddPackageViewModel()
+        let viewModel = WooShippingAddCustomPackageViewModel()
 
         // When
         viewModel.clearFieldValues()
@@ -28,7 +31,7 @@ final class WooShippingAddPackageViewModelTests: XCTestCase {
 
     func test_it_with_not_all_field_values_set() {
         // Given
-        let viewModel = WooShippingAddPackageViewModel()
+        let viewModel = WooShippingAddCustomPackageViewModel()
 
         // When
         viewModel.clearFieldValues()
@@ -41,11 +44,11 @@ final class WooShippingAddPackageViewModelTests: XCTestCase {
 
     func test_it_with_all_field_values_set() {
         // Given
-        let viewModel = WooShippingAddPackageViewModel()
+        let viewModel = WooShippingAddCustomPackageViewModel()
 
         // When
         viewModel.clearFieldValues()
-        for dimensionType in WooShippingAddPackageDimensionType.allCases {
+        for dimensionType in WooShippingPackageDimensionType.allCases {
             viewModel.fieldValues[dimensionType] = "1"
         }
